Respond to CR - remove iceland/norway regions in favor of three-letter country codes for now.
- Add an alias for PrivacyNoticeTemplate + PrivacyNotice
- Relocate test now that method has moved
